Health and money bar updates, projectile optimization
Added a changing healthbar and moneybar. The way they are updated in the
GUI is suboptimal, but will do for now since it's effects are minimal.
I'm not sure whether you should have acces to the GUI from the World at
all. Also added money on monster death. Projectile optimization was
about deleting projectiles that are offscreen. It was important to do
this because of the way projectiles kill monsters by going through a
collection of active monsters. Every extra projectile therefore adds
extra calculations.