Game logic and Project structure
Added new subclasses of monster. I need to decide what is the best way
to draw them. Maybe there could be some useful Scala syntax to use.
Projectiles now decrease hp instead of killing instantly. Updated the
structure of the project so that it can be by just downloading the git
project. Updated the GUI a bit.