Bug resolved - partially
projectiles works with ListBuffer, but not with Buffer. In reality
changing the collection type probably does not resolve the cause of the
bug, or maybe it's due to differences in linked list vs array data
structure. If it's about structure then I need to do more research to
understand if it's a bug I probably still need to do more research to
find what caused the null-pointer exception. 

Anyhow for time being it's better to move forward due to time pressures.
This was in the end a pretty good challenge in debugging and using
version control to find out what was wrong.