Undo
Undid the changes regarding the grid. For now I'll use it only for
drawing. Changing the Path would require a nearly complete rework of the
logic. For future updates after the DL I might consider writing the code
again from scratch where the path class is replaced by a sequence of
grid tiles. But for now I'll have to stick with path consisting of Pos.
This will affect the ability of user to interact with the project, but
also greatly increase this weekends wellbeing.