Tower drawing and balancing, new level
Added drawing for towers. Balanced their stats so that it is now good to
also buy basic towers. Added a new level which is passable but fairly
hard. The gameplay is fun considering the game has only 3 monsters and 3
towers. Found a bug that the program doesn't crash if the first file is
corrupted. No time to fix it before submit, though it should be okay if
the user doesn't modify the testFile.